requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
click==8.1.3
discord-py-interactions==4.4.0
discord-typings==0.5.1
dnspython==2.3.0
emoji==2.2.0
exceptiongroup==1.1.1
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
pymongo==4.3.3
pytest==7.3.0
python-dotenv==1.0.0
tomli==2.0.1
typing_extensions==4.5.0
yarl==1.8.2
